Improve training correctness, checkpoint integrity, and alignment pipeline
  - Rewrite MDLM compute_loss as faithful continuous-time NELBO with
    schedule-derived weighting (-α'(t)/(1-α(t))) via central difference
  - Fix LR scheduler off-by-one: first optimizer step now uses intended
    warmup LR instead of zero
  - Add checkpoint identity binding: run_meta.json (signature + config
    snapshot) validated on resume; model_type.txt prevents cross-family
    loading (e.g. D3PM checkpoint loaded as MDLM)
  - Trainer-specific critical-field validation: DPO asserts dpo_beta,
    GRPO asserts all grpo_* hyperparameters match on resume
  - Require explicit frozen reference model for DPO/GRPO instead of
    deepcopy (which silently drifts on resume); persist ref_path.txt
  - Add GRPO trainer with PPO-style clipped surrogate, inner epochs,
    group-relative advantages, and KL penalty against frozen reference
  - Add GSM8K task module, generation utilities, diffusion samplers
  - Faithful RNG resume: snapshot loader generator state at epoch
    boundaries, fast-forward on restore
  - Update all scripts (pretrain, SFT, DPO, GRPO, compare_*, sample,
    evaluate, generate) for new trainer signatures

+++ minilab/nn/norm.py
@@ -14,7 +14,7 @@ class RMSNorm(nn.Module):
 
     def forward(self, x):
         rms = torch.rsqrt(x.float().pow(2).mean(-1, keepdim=True) + self.eps)
-        return (x * rms).to(x.dtype) * self.weight
+        return (x * rms).to(x.dtype) * self.weight.to(x.dtype)
 
 
 @register_norm("layernorm")
